Refactored the post processing code into a generator function
I intended this to be used in the main function, where the main function
would have a collection of post processors that it would give the
yielded post to. However, I'm not sure if that's the right design.
Instead I think I should stick to the original plan and just change the
after() method for each class that I want to enable.